Update getRetainedSize methods to return long

+++ src/main/java/io/airlift/slice/DynamicSliceOutput.java
@@ -61,7 +61,7 @@ public class DynamicSliceOutput
     }
 
     @Override
-    public int getRetainedSize()
+    public long getRetainedSize()
     {
         return slice.getRetainedSize() + INSTANCE_SIZE;
     }
