+++ src/main/java/io/airlift/slice/OutputStreamSliceOutput.java
@@ -102,7 +102,7 @@ public class OutputStreamSliceOutput
     }
 
     @Override
-    public int getRetainedSize()
+    public long getRetainedSize()
     {
         return slice.getRetainedSize() + INSTANCE_SIZE;
     }
